+++ issue-repro-compose/src/main/java/com/airbnb/lottie/issues/compose/ComposeIssueReproActivity.kt
@@ -7,13 +7,10 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.material.Text
 import androidx.compose.material.TextButton
 import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
 import com.airbnb.lottie.compose.*
-import kotlinx.coroutines.NonCancellable
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.isActive
-import kotlinx.coroutines.withContext
 
 class ComposeIssueReproActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -25,53 +22,50 @@ class ComposeIssueReproActivity : AppCompatActivity() {
 
     @Composable
     fun Content() {
-        val states = remember { MutableStateFlow(0) }
-        val state by states.collectAsState()
-        val progress = remember { mutableStateOf(0f) }
-
+        var state by remember { mutableStateOf(0) }
         val compositionResult = rememberLottieCompositionResult(LottieCompositionSpec.RawRes(R.raw.loading))
 
-        LaunchedEffect(Unit) {
+        val progress by lottieTransition(state) { progress ->
             val composition = compositionResult.await()
-            states.collectLatest { state ->
-                when (state) {
-                    0 -> {
+            when (state) {
+                0 -> {
+                    animateLottieComposition(
+                        composition,
+                        progress,
+                        clipSpec = LottieAnimationClipSpec.MinAndMaxProgress(0f, 0.25f),
+                        cancellationBehavior = LottieCancellationBehavior.AtEnd,
+                    )
+                }
+                1 -> {
+                    while (isActive) {
                         animateLottieComposition(
                             composition,
                             progress,
-                            clipSpec = LottieAnimationClipSpec.MinAndMaxProgress(0f, 0.25f),
-                            cancellationBehavior = LottieCancellationBehavior.AtEnd,
-                            )
-                    }
-                    1 -> {
-                        while (isActive) {
-                            animateLottieComposition(
-                                composition,
-                                progress,
-                                clipSpec = LottieAnimationClipSpec.MinAndMaxProgress(0.25f, 0.75f),
-                                cancellationBehavior = LottieCancellationBehavior.AtEnd,
-                            )
-                        }
-                    }
-                    else -> {
-                        animateLottieComposition(
-                            composition,
-                            progress,
-                            clipSpec = LottieAnimationClipSpec.MinAndMaxProgress(0.75f, 1f),
+                            clipSpec = LottieAnimationClipSpec.MinAndMaxProgress(0.25f, 0.75f),
                             cancellationBehavior = LottieCancellationBehavior.AtEnd,
                         )
                     }
                 }
+                else -> {
+                    animateLottieComposition(
+                        composition,
+                        progress,
+                        clipSpec = LottieAnimationClipSpec.MinAndMaxProgress(0.75f, 1f),
+                        cancellationBehavior = LottieCancellationBehavior.AtEnd,
+                    )
+                }
             }
         }
 
-        Column {
+        Column(
+            horizontalAlignment = Alignment.CenterHorizontally,
+        ) {
             LottieAnimation(
                 compositionResult(),
-                progress.value,
+                progress,
             )
             TextButton(
-                onClick = { states.value = (state + 1) % 3 },
+                onClick = { state = (state + 1) % 3 },
             ) {
                 Text(state.toString())
             }
