+++ lottie-compose/src/main/java/com/airbnb/lottie/compose/lottieTransition.kt
@@ -1,23 +1,26 @@
 package com.airbnb.lottie.compose
 
 import androidx.compose.runtime.*
-import com.airbnb.lottie.LottieComposition
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.flow.collectLatest
 
 @Composable
 fun <T> lottieTransition(
-    composition: LottieComposition?,
     state: T,
-    animate: suspend (composition: LottieComposition?, progress: MutableState<Float>) -> Unit,
-) {
+    animate: suspend CoroutineScope.(progress: MutableState<Float>) -> Unit,
+): State<Float> {
     val progress = remember { mutableStateOf(0f) }
     val states = remember { MutableStateFlow(state) }
     states.value  = state
 
     LaunchedEffect(Unit) {
-        states.collect {
-            animate(composition, progress)
+        states.collectLatest {
+            coroutineScope {
+                animate(progress)
+            }
         }
     }
+    return progress
 }
